requirements.txt
slack_sdk
google-generativeai
requests